Rewrite Soniox integration to match actual API
BREAKING: Complete rewrite of the Soniox integration layer.

The old code called a non-existent REST endpoint (/v1/sessions) which
returned 404. The actual Soniox real-time API is WebSocket-only:

Architecture (before - broken):
  Browser -> Backend -> Soniox REST /v1/sessions (404!)

Architecture (after - correct):
  Browser -> Backend POST /api/soniox/temporary-key
  Backend -> Soniox REST /v1/auth/temporary-api-key
  Browser -> Soniox WebSocket wss://stt-rt.soniox.com/transcribe-websocket

Changes:

Backend:
- Replace session CRUD routes with single /temporary-key endpoint
- Backend now only issues temporary API keys (keeps real key secure)
- Fix CORS: parse comma-separated origins into an array
- Simplify sonioxService to single createTemporaryApiKey function

Frontend soniox.ts:
- Connect directly to wss://stt-rt.soniox.com/transcribe-websocket
- Send JSON config as first message (api_key, model, translation, etc.)
- Send raw PCM Int16 binary frames (not base64 JSON)
- Parse token-based responses (text, is_final, translation_status)
- Configure one-way translation (es -> en) with church context

Frontend audioCapture.ts:
- Output Int16Array PCM at 16kHz (Soniox requirement)
- Request 16kHz sample rate from AudioContext
- Convert Float32 to Int16 in ScriptProcessor callback

Frontend hooks:
- Rewrite useSonioxSession with connect/sendAudio/finish/disconnect
- Track TranscriptEntry[] with type 'transcription' | 'translation'
- Update useAudioCapture callback type to Int16Array

Frontend components:
- Rewrite LiveFeed to show transcription + translation panels
- Rewrite MicrophoneSetup for new hook API
- Simplify App.tsx (remove unused store dependencies)
- Remove unused TranslationDisplay component dependency

Config:
- Remove unused path aliases from tsconfig.json
- Install @types/cors for backend type safety

+++ src/hooks/useAudioCapture.ts
@@ -42,8 +42,8 @@ export const useAudioCapture = () => {
     }
   }, []);
 
-  // Start capturing audio
-  const startCapture = useCallback((audioCallback: (data: Float32Array) => void) => {
+  // Start capturing audio (callback receives Int16 PCM)
+  const startCapture = useCallback((audioCallback: (data: Int16Array) => void) => {
     try {
       if (!isInitialized) {
         throw new Error('Audio capture not initialized');
